Remove jline color-coded console (10 MB -> 9.8 MB)
JLine was used to provide a prompt with tab completion and ANSI color
output, however, to run the server in the background such as on Heroku
in a daemon, it has to be disabled otherwise it will spam ">>>>"
everywhere, thinking there is more user input at the prompt.

This could already be disabled in the config with use-jline: false, as I
have done here:
https://github.com/satoshinm/heroku-glowstone/commit/4a7e828d5d865490698d82ed750a777df32d0a4a
but removing it saves 0.2 MB and the need to make this change.