Minimize the jar but always include Glowkit 9.8 MB -> 6.3 MB
Removes a large amount of code, which wasn't referenced anywhere. Note
this may break plugins (not tested extensively), but to reduce the
chance of breakage I add an include for the Glowkit API, so anything
using only that should be safe!

This works well enough to run https://github.com/satoshinm/WebSandboxMC